S5PC11X: SENSOR: Change LightSensor to return lux
Our initial LightSensor HAL returned the raw adc from the driver.
Now we implement an initial conversion from adc to lux levels.
It's not perfect, but lux varies quite a bit anyway and the levels
are so large that precision isn't generally as important as being
in the right ballpark.  We can ask HW team to do get more precision
if necessary.

Change-Id: I7ad09bb13831b8100c255ab9cd7287a4b684619d
Signed-off-by: Mike J. Chen <mjchen@sta.samsung.com>