SP5C11X: libcamera: Improve start up time
This is part of change to the interface between driver
and HAL so that driver doesn't do anything if a setting
is the same as current.  The HAL part is to send all
settings for preview in one ioctl, instead of about
a dozen to save even more time (total is about 500ms).

Change-Id: I46392e970fd1725719359cdf7fb6ca2006beba4c
Signed-off-by: Mike J. Chen <mjchen@sta.samsung.com>