S5PC11X: libcamera: Fix CTS issue and update focus distance
Initialize the mSecCamera settings in initDefaultParameters in case
setParameter() isn't called by the application (it's not required).
Also had to bump the delay in SecCamera::DeinitCamera() or else
the next CTS test was not able to start because SecCamera::initCamera
would fail to open /dev/video0 because it hadn't been completely
released by the previous camera test.

Also updated focus distances for back camera with new info from TechWin.

Change-Id: Ib84890a466e5f96908a333423e4dcf146a744252
Signed-off-by: Mike J. Chen <mjchen@sta.samsung.com>