Issue 3058745: fix audio HAL downsampler.
Latest audio HAL downsampler implementation had a problem that
caused one sample to be dropped if the input buffer size was odd
while downsampling by a factor of 2.
This explains why record timing was correct for 22 and 16kHz but
not for 11 or 8kHz: in the later case, the second stage of resampling
(22 to 11) receives 503 frames for each buffer.

Change-Id: Ib8fcba3ddfbbab50a908e6b0a6cdc2b398acd862