Fix issue 3123882.
When in call and an incoming call is received, the call mode is temporarily
switched back to ringtone before the new call is answered. This was deconfiguring
in call audio path and we did not recover properly when transitioning back from
ring tone to in call.

The fix consists in not changing the in call audio config when changing from
in call to ring tone but only when going back to normal mode.

Change-Id: I78928e4a0043f114accf1f602f255321161c45ad