Fix issue 3124837.
There is a self deadlock condition in AudioStreamOutALSA::write() when a error occurs
writing to the driver. In this case, the mutex is held and we call standby() which tries
to acquire the mutex too.
The same condition is fixed also in AudioStreamInALSA::read()

Additional changes:
- Use a single instance of the alsa mixer to avoid simultaneous route change requests
to the kernel driver.
- Protect all route and mode changes by common mutex to avoid changing routes while
reading or writing from/to the kernel driver.
- Merged support for headphones (w/o mic).
- Code cleanup

Change-Id: I62bfe34f4dd4538e410f030bd0063c8835b5ee54