Fix issue 3130388.
The problem comes from the way the codec driver handles codec path and clock configuration
when a call is terminated: the driver relies on the pcm output stream to be stopped and restarted
between two calls. This is not the case if music is playing when a first call is received and not stopped
before a second call is received.

Implemented a workaround in audio HAL that forces the output stream into standby when the call is terminated.

Change-Id: I58ed6f4e3a5edb0e363c06580e2b97ce68988c4c