libcamera: don't invoke callback when startPreview fails
startPreview() of SecCameraHWInterface.cpp should not invoke a
callback when startPreview fails. It will cause a deadlock.
CameraService was holding mLock in any method call from applications.
In those calls, camera HAL must not invoke a callback and it also
needs to acquire the mLock. startPreview should just return an error
and apps will get an exception.

Change-Id: I89f3a614a0efa56abcdde17ba5dd5f7a5e54911c