Fix issue 3148223.
This change fixes issue 3148223 but also other issues with audio routing
that could happen from time to time. The cause for those issues is the
weakness of kernel codec driver with regard to timing and order in which audio
path and pcm stream state changes are requested.
Other issues reported are:
- music sometimes distorted/inaudible after several voice search / phone calls
- loss of incall audio after switching between SIP and PSTN Calls.

The fix consists in systematically make sure that the codec driver is shutdown
whenever an audio path change is requested: both output and input streams are forced
into standby every time a change occurs in phone state, input path selection or
output path selection.

Change-Id: I3a534393cf6e551ce7b9f7569bcbcf33271e6a6c