Fix issue 3161227.
When a call is answered, the audio path is configured two times:
- a first time by audio HAL when changing audio mode from RINGTONE to IN_CALL, where
current device is applied to voice call path as a default configuration
- a second time when audio policy manager defines which device should be used. This happens
several hundred ms later so that the ringtone can be stopped before audio is routed to the
earpiece.
The uplink audio disruption apparently occurs because we switch the incall path first to
speaker(as this is the current path for ringtone) and then to ear piece.

The fix consists in removing the first audio path configuration by audio HAL when entering
in call which is in fact redundant with the configuration done by audio policy manager.

Change-Id: Ib99674dd3dcbba33e3d766e311ac54e919e64660