Issue 3137532: removed garbage at start of capture
The first call to pcm_read() just starts PCM capture and does not return any
data although the read status is valid. Data present in memory where the read
buffer was allocated is fed to the resampler for the first audio buffer.

The fix consists in modifying pcm_read() to return valid data from the first read.

Change-Id: I489e398db570e709d1f64507b67a433d6a36c3d1