Fix issue 3201189
The issue comes from the fact that acquiring a mutex can take upto several
seconds if the acquisition is attempted by a low priority thread and a high priority
thread is frequently acquiring and releasing the same mutex.
This happens when setMode() tries to lock the output stream mutex while playback is active.

The fix is more a workaround consisting in boosting the priority of the thread executing
setMode() until the mutex is acquired.

Change-Id: I0921c215759dcd86d6ffb03669b5f94691878ffb