Fix issue 3305305.
The problem is that when the voice search tone is started, audio capture
is still active and the output stream write function needs to place
the input stream in standby to reconfigure the kernel driver.
To do so it has to acquire the input stream mutex but as the input stream thread
holds the lock most of the time while sleeping in the driver waiting for more data,
this is very difficult and can take several seconds.

The fix consists in forcing a sleep in the next read() when another function needs
to acquire the input stream lock.

The same change is done for output stream write() function.

Also removed the workaround for issue 3201189 in setMode() (thread priority bump) as this
change addresses the same problem.

Change-Id: I3a5e672717752f83dfedce822a18748b165b0a5a