S5PC11X: SENSOR: The status of orientation sensor show correcly when it is '0' (2)
When the status goes to '0', that event is ignored
because its event type is 'REL' which filters out number '0'.
So if the real status changes into '0' from '1, 2, 3',
the status on the display still remains '1, 2, 3' even though actual status is '0'.
Getting around this problem by reporting event as 4 instead of '0'.
If sensor HAL got '4', it retansforms it into '0'.

Change-Id: I41a92be60c0b0a96ac8dc9fba4fd4d21d8abc76c
Signed-off-by: tim.sk.lee <tim.sk.lee@samsung.com>