Fix issue 4126225: setMicMute() execution time
The cause is that the thread executing setMicMute() can wait for
several seconds before acquiring the input thread mutex when audio
capture is active.
The fix consists in forcing a sleep of the capture thread when
standby() (called by setMicMute()) wants the mutex.

Applied the same fix to output stream and to setParameters() function.

Change-Id: I3e55670d2aad16b67d44ca8582ed16398143ff6e