Video decode component support gralloc color format.
That returns the current color format is a 0x100.
But, It is not the normal way In terms of omx. (It is spec out)
OMX spec is OMX_COLOR_FormatVendorStartUnused = 0x7F000000, /**< Reserved region for introducing Vendor Extensions */

I think.... I hope the color format is the value 0x7F000100.
stagefright use the value(0x7F000100) after masking.

ex)
    err = native_window_set_buffers_geometry(
            mNativeWindow.get(),
            def.format.video.nFrameWidth,
            def.format.video.nFrameHeight,
            (def.format.video.eColorFormat & 0x1FF));

Change-Id: If0578eb35338ff9bd8d0cf8cd8f1435af92aba04
Signed-off-by: SeungBeom Kim <sbcrux.kim@samsung.com>