Remove software egl
Since some apps are not correct running it on crespo,
like this goes faster and better.

In example:
Quadrant: Planet demo groe from 26fps to 56fps, and removes this extrange
bad effect on rendering the little plannet.
OpeGL3d Renders more fps and better quality
In general most 3D games are faster with this removed

Change-Id: I7fccbccb96e1ccafe9cf3266cc7c8f0fb641cd48