libcamera: record GPS coordinates with greater precision.
Record latitude and longitude to 7 decimal places (~5 cm) and altitude
to 1 cm. The previous behaviour was to record lat/long to 1 arcsecond
(~15 m), which was rounded down, so the potential error was up to ~30 m.

Addresses AOSP issue 16626.

Original author: Vernon Tang <vt@foilhead.net>
Suggested by: Wu-cheng Li
Reviewed by: jubeam@samsung.com

Change-Id: I031858972b939f82f40c0f7299d8eff48bc4f4c5
related-to-bug: 5550852