Fix regression in bugreport key combo
In Gingerbread, pressing the bugreport key combo when USB debugging was
enabled created a bugreport. This key combo was changed to run
bugmailer.sh in Honeycomb. This was further changed to run some
different things in ICS, but it was disabled for user builds.

These changes inadvertently made bugreport key combo not work on user
builds. When the combo is pressed, this appears in dmesg:

init: cannot find '/system/bin/bugmailer.sh', disabling 'bugreport'

Change-Id: If3a09d9dabdc598bf714512bc402bdcb8767d25c