CrespoParts: Revise English translations to bring in-line with conventions
This patch brings consistency to the formatting of the strings with the
rest of the system.

Headers should be capitalized automatically, we only capitalize the first
letter of each string, we don't put periods at the end of description
strings, and so forth.

Ported from CM10, also includes some revision as per Kim Lê's original
suggestions.

Change-Id: Ia8827c93851d032792832efba33452439f5210e8